requirements.txt
Flask==2.3.2
openai==0.27.7
pdfminer.six==20221105
python-docx==0.8.11
python-pptx==0.6.21